renamed some things, cleaned up logs and comments, moved some things
