+++ packages/gamut-labs/src/PlanCard/consts.ts
@@ -1,5 +1,3 @@
-import { Currency, CurrencySymbol } from './types';
-
 interface Detail {
   id: string;
   title: string;
@@ -62,27 +60,3 @@ export const planDetails: PlanDetails = {
     },
   },
 };
-
-const currencySymbols: Record<Currency, CurrencySymbol> = {
-  [Currency.USD]: CurrencySymbol.DOLLAR,
-  [Currency.CAD]: CurrencySymbol.DOLLAR,
-  [Currency.AUD]: CurrencySymbol.DOLLAR,
-  [Currency.EUR]: CurrencySymbol.EURO,
-  [Currency.GBP]: CurrencySymbol.POUND,
-  [Currency.INR]: CurrencySymbol.RUPEE,
-  [Currency.SEK]: CurrencySymbol.KRONA,
-  [Currency.NOK]: CurrencySymbol.KRONE,
-  [Currency.PLN]: CurrencySymbol.ZLOTY,
-  [Currency.DKK]: CurrencySymbol.KRONE,
-  [Currency.BRL]: CurrencySymbol.REAL,
-  [Currency.BGN]: CurrencySymbol.LEV,
-  [Currency.CHF]: CurrencySymbol.FRANC,
-  [Currency.CZK]: CurrencySymbol.KORUNA,
-  [Currency.HRK]: CurrencySymbol.KUNA,
-  [Currency.HUF]: CurrencySymbol.FORINT,
-  [Currency.MXN]: CurrencySymbol.PESO,
-  [Currency.RON]: CurrencySymbol.LEU,
-};
-
-export const getCurrencySymbol = (currency: Currency) =>
-  currencySymbols[currency];
